% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele.balance.filter.R
\name{filter.allele.balance}
\alias{filter.allele.balance}
\title{Filter out heterozygous sites failing an allele balance check}
\usage{
filter.allele.balance(vcfR)
}
\arguments{
\item{vcfR}{a vcfR object}
}
\value{
The vcfR object input, with the sites failing the allele balance filter converted to 'NA'
}
\description{
This function requires a vcfR object as input, and returns a vcfR object filtered
to convert heterozygous sites with an allele balance falling outside of .25-.75 to 'NA'.
From: https://www.ddocent.com/filtering/ "Allele balance: a number between 0 and 1
representing the ratio of reads showing the reference allele to all reads,
considering only reads from individuals called as heterozygous, we expect that the
allele balance in our data (for real loci) should be close to 0.5".
}
