% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_depth.R
\name{vis_depth}
\alias{vis_depth}
\title{Visualize the effect of varying the m parameter on the mean depth of each sample}
\usage{
vis_depth(output = NULL)
}
\arguments{
\item{output}{A list containing 5 dataframes generated by optimize_m()}
}
\value{
A plot showing the depth of each sample at each given m value
}
\description{
This function takes the list of dataframes output by optimize_m() as input.
The function then uses ggplot2 to visualize the effect of m on depth.
}
\examples{
vis_depth(output =
readRDS(system.file("extdata","optimize.m.output.RDS",package="RADstackshelpR",mustWork=TRUE)))
}
